shell

+++ hello_world/6-size.c
@@ -10,6 +10,6 @@ printf("%lu\n", sizeof(char));
 printf("%lu\n", sizeof(int));
 printf("%lu\n", sizeof(long int));
 printf("%lu\n", sizeof(long long int));
-printf("%lu", sizeof(double));
+printf("%lu", sizeof(float));
 return (0);
 }
